changed several values to doubles for improved accuracy, fix issue with envelope attack

+++ src/synth/AudioGraph.cs
@@ -395,7 +395,7 @@ namespace Synth
         //     }
         // }
 
-        public void Process(float increment)
+        public void Process(double increment)
         {
             lock (_lock)
             {
