+++ src/synth/ReverbEffectNode.cs
@@ -18,7 +18,7 @@ namespace Synth
             reverbModel = new ReverbModel();
         }
 
-        public override void Process(float increment)
+        public override void Process(double increment)
         {
             var inputs = GetParameterNodes(AudioParam.StereoInput);
             if (inputs == null || inputs.Count == 0)
